Graceful MQTT Connection Loss Handling & Added Reset to AP Mode
Created a variable that tracks the number of connection attempts.  If the device fails connect to the MQTT server after this many attempts it will stop checking for MQTT.  This should prevent the system from hanging up when the there is no MQTT to connect to. 

Added a button the diagnostics page to reset the wifi connection data and return to AP Mode.  This will wipe your credentials and return the device to AP mode.  You'll need to log back into the WiFi manager and set your credentials again.  

Known Issue:  On reset of the device, the MQTT parameters are lost (and are not pulled from the saved memory).  Workaround is to set these parameters correctly in the code initially so that the system defaults are correct.